refactorizado de game state
